passing features with delete button removed for editor role

+++ app/helpers/application_helper.rb
@@ -20,7 +20,7 @@ module ApplicationHelper
         ret += '<em>'
         ret += item.updated_at.strftime("Last updated %B %d @ %H:%M by #{item.updated_by.puid}")
         ret += '</em>'
-        ret += button_to "delete", admin_page_path(item), :method => :delete, :confirm => "Are you sure you want to delete the page #{item.title}" 
+        ret += button_to "delete", admin_page_path(item), :method => :delete, :confirm => "Are you sure you want to delete the page #{item.title}" if can? :destroy, item
         ret += '</div>'
         ret += tree_ul(item.children, false, &block) if item.children.size > 0
         ret += '</li>'
